requirements.txt
requests
pandas
python-chess
matplotlib
streamlit
plotly